requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.32
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
boto3==1.34.162
Werkzeug==3.0.3
